improved ui and data visibility

+++ view/templates/viewIndicators.blade.php
@@ -69,7 +69,7 @@
 
 <main id="main" class="main">
 
-  <div class="pagetitle">
+  <div class="pagetitle mt-3">
     <h1>Indicators</h1>
     <nav>
       <ol class="breadcrumb">
@@ -146,15 +146,17 @@
                         Select Action
                       </button>
                       <div class="dropdown-menu" aria-labelledby="actionDropdown">
-                        @if($role == 'Administrator')
-                        @if($indicator['status'] == 'draft' || $indicator['status'] == 'review')
-                        <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Editing an indicator is only available when the indicator is in draft or review state." href="/{{$appName}}/dashboard/indicators/edit?id={{$indicator['id']}}" class="dropdown-item">
-                          <i class="bi bi-pencil-square"></i> Edit Indicator
-                        </a>
-                        @endif
                         <a href="/{{$appName}}/dashboard/indicators/responses/all?id={{$indicator['id']}}" class="dropdown-item">
                           <i class="bi bi-eye"></i> View Indicator Responses
                         </a>
+                        @if($role == 'Administrator')
+                          @if($myOrganisation['id'] == $indicator['organization_id'] || $myOrganisation['name'] == 'Administrator')
+                            @if($indicator['status'] == 'draft' || $indicator['status'] == 'review')
+                            <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Editing an indicator is only available when the indicator is in draft or review state." href="/{{$appName}}/dashboard/indicators/edit?id={{$indicator['id']}}" class="dropdown-item">
+                              <i class="bi bi-pencil-square"></i> Edit Indicator
+                            </a>
+                            @endif
+                          @endif
                         @endif
                         @if($role == 'Viewer')
                         <a href="/{{$appName}}/dashboard/indicators/responses/all?id={{$indicator['id']}}" class="dropdown-item">
@@ -162,33 +164,38 @@
                         </a>
                         @endif
                         @if($role == 'User' || $role == 'Administrator')
-                        @if($indicator['status'] == 'draft')
-                        <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Adding responses to an indicator is only available when the indicator is in draft or review state." href="/{{$appName}}/dashboard/indicators/responses/add?id={{$indicator['id']}}" class="dropdown-item">
-                          <i class="bi bi-plus-circle"></i> Add Response
-                        </a>
+                          @if($myOrganisation['id'] == $indicator['organization_id'] || $myOrganisation['name'] == 'Administrator')
+                            @if($indicator['status'] == 'draft')
+                            <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Adding responses to an indicator is only available when the indicator is in draft or review state." href="/{{$appName}}/dashboard/indicators/responses/add?id={{$indicator['id']}}" class="dropdown-item">
+                              <i class="bi bi-plus-circle"></i> Add Response
+                            </a>
+                            @endif
+                            @if($indicator['status'] == 'draft')
+                            <a id="review-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=review" class="dropdown-item">
+                              <i class="bi bi-pencil"></i> Review Indicator
+                            </a>
+                            @endif
+                            @if($indicator['status'] == 'review')
+                            <a id="publish-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=public" class="dropdown-item">
+                              <i class="bi bi-globe"></i> Make Indicator Public
+                            </a>
+                            @endif
+                            @if($indicator['status'] == 'review' || $indicator['status'] == 'public')
+                            <a id="archive-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=archived" class="dropdown-item">
+                              <i class="bi bi-archive"></i> Archive Indicator
+                            </a>
+                            @endif
+                          @endif  
                         @endif
-                        @if($indicator['status'] == 'draft')
-                        <a id="review-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=review" class="dropdown-item">
-                          <i class="bi bi-pencil"></i> Review Indicator
-                        </a>
-                        @endif
-                        @if($indicator['status'] == 'review')
-                        <a id="publish-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=public" class="dropdown-item">
-                          <i class="bi bi-globe"></i> Make Indicator Public
-                        </a>
-                        @endif
-                        @if($indicator['status'] == 'review' || $indicator['status'] == 'public')
-                        <a id="archive-indicator-btn" href="/{{$appName}}/dashboard/indicators/status/update?id={{$indicator['id']}}&status=archived" class="dropdown-item">
-                          <i class="bi bi-archive"></i> Archive Indicator
-                        </a>
-                        @endif
-                        @endif
+
                         @if($role == 'Administrator')
-                        @if($indicator['status'] == 'draft')
-                        <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Deleting an indicator is only available when the indicator is in draft state." href="/{{$appName}}/dashboard/manage-indicators/delete/?id={{$indicator['id']}}" class="dropdown-item text-danger" id="delete-btn">
-                          <i class="bi bi-trash"></i> Delete Indicator
-                        </a>
-                        @endif
+                          @if($myOrganisation['id'] == $indicator['organization_id'] || $myOrganisation['name'] == 'Administrator')
+                            @if($indicator['status'] == 'draft')
+                            <a data-bs-toggle="tooltip" data-bs-placement="left" data-bs-original-title="Deleting an indicator is only available when the indicator is in draft state." href="/{{$appName}}/dashboard/manage-indicators/delete/?id={{$indicator['id']}}" class="dropdown-item text-danger" id="delete-btn">
+                              <i class="bi bi-trash"></i> Delete Indicator
+                            </a>
+                            @endif
+                          @endif
                         @endif
                       </div>
                     </div>
